ch15/requirements.txt
asgiref==3.7.2
Django==4.1.6
django-registration-redux==2.12
django-uuid-upload-path==1.0.0
Pillow==9.4.0
pip==22.3.1
setuptools==65.5.1
sqlparse==0.4.4
typing_extensions==4.8.0
wheel==0.41.2
